refactor: Merge types package to core
