+++ ecal/publisher/publisher_test.go
@@ -4,14 +4,14 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/DownerCase/ecal-go/ecal"
 	"github.com/DownerCase/ecal-go/ecal/publisher"
-	"github.com/DownerCase/ecal-go/ecal/types"
 	testutilpublisher "github.com/DownerCase/ecal-go/internal/ecaltest/testutil_publisher"
 )
 
 func TestNewPublishers(t *testing.T) {
 	for i := range 100 {
-		ptr, err := publisher.New(fmt.Sprintf("testPubTopic-%v", i), types.DataType{})
+		ptr, err := publisher.New(fmt.Sprintf("testPubTopic-%v", i), ecal.DataType{})
 		if err != nil {
 			t.Error(err)
 		}
